search in compressed array

+++ listOps.h
@@ -11,4 +11,4 @@ using namespace std;
 
 int open(string word, unordered_map<string,lexiconData> &lexicon, List *list);
 int startSearch(unordered_map<string,lexiconData> &lexicon);
-int nextGEQ(List *list, uint64_t k, int *docID);
+int nextGEQ(List *list, uint64_t k, uint64_t *docID);
